+++ changelog.md
@@ -1,5 +1,12 @@
 # SmartQL changelog
 
+## 13.0.0
+
+### Major
+
+- [smartql.mjs](./smartql.mjs) now uses one object argument.
+- Removed main index export for [smartql.mjs](./smartql.mjs)
+
 ## 12.0.0
 
 ### Major